+++ src/Help/index.js
@@ -1,29 +1,36 @@
-import React from 'react'
-import propTypes from 'prop-types'
 import cx from 'classnames'
+import propTypes from 'prop-types'
+import React from 'react'
 
-import { theme, spacers } from '../theme.js'
+import { statusPropType } from '../common-prop-types'
+import { spacers, theme } from '../theme.js'
 
-const Help = ({ children, valid, error, warning, className }) => (
+const Help = ({ children, valid, error, warning, className, indent }) => (
     <p
         className={cx(className, {
             valid,
             error,
             warning,
+            indent,
         })}
     >
         {children}
 
         <style jsx>{`
             p {
-                height: 12px;
-                padding-top: ${spacers.dp8};
-                padding-left: ${spacers.dp16};
                 font-size: 12px;
-                margin: 0;
                 line-height: 12px;
-                cursor: help;
+                margin: 0;
                 color: ${theme.default};
+                white-space: nowrap;
+                overflow: hidden;
+                text-overflow: ellipsis;
+                cursor: help;
+                padding-top: ${spacers.dp8};
+            }
+
+            .indent {
+                padding-left: ${spacers.dp16};
             }
 
             .valid {
@@ -41,14 +48,17 @@ const Help = ({ children, valid, error, warning, className }) => (
     </p>
 )
 
-Help.defaultProps = {}
-
 Help.propTypes = {
     className: propTypes.string,
     children: propTypes.string.isRequired,
-    error: propTypes.bool,
-    valid: propTypes.bool,
-    warning: propTypes.bool,
+    error: statusPropType,
+    valid: statusPropType,
+    warning: statusPropType,
+    indent: propTypes.bool,
+}
+
+Help.defaultProps = {
+    indent: true,
 }
 
 export { Help }