+++ src/MenuItem/styles.js
@@ -3,13 +3,11 @@ import { colors, spacers } from '../theme.js'
 
 export default css`
     li {
-        display: flex;
         position: relative;
-        flex-direction: row;
-        align-items: center;
         height: 48px;
-        padding: 0 ${spacers.dp24};
+        padding: 0;
         cursor: pointer;
+        list-style: none;
     }
 
     li div.label:not(:first-child) {
@@ -23,28 +21,6 @@ export default css`
     li:active,
     li.active {
         background-color: ${colors.grey400};
-    }
-
-    .label {
-        color: ${colors.grey900};
-        font-size: 15px;
-
-        white-space: nowrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-
-        user-select: none;
-        padding-right: ${spacers.dp12};
-    }
-
-    .icon {
-        box-sizing: border-box;
-
-        margin-right: ${spacers.dp16};
-        color: #404040;
-        font-size: 24px;
-        pointer-events: none;
-        user-select: none;
     }
 
     .dense {
@@ -69,5 +45,37 @@ export default css`
     .disabled .icon,
     .disabled .label {
         color: rgba(0, 0, 0, 0.3);
+    }
+
+    .link {
+        display: block;
+        height: 100%;
+        padding: 0 ${spacers.dp24};
+        text-decoration: none;
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+    }
+
+    .label {
+        color: ${colors.grey900};
+        font-size: 15px;
+
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+
+        user-select: none;
+        padding-right: ${spacers.dp12};
+    }
+
+    .icon {
+        box-sizing: border-box;
+
+        margin-right: ${spacers.dp16};
+        color: #404040;
+        font-size: 24px;
+        pointer-events: none;
+        user-select: none;
     }
 `