+++ src/SelectField/index.js
@@ -1,8 +1,9 @@
 import propTypes from 'prop-types'
 import React from 'react'
 
-import { ArrowDown } from '../icons/Arrow'
+import { statusPropType } from '../common-prop-types'
 import { LabelFilled, LabelOutlined } from '../FieldLabel'
+import { ArrowDown } from '../icons/Arrow'
 import { Select } from './Select'
 
 const TailIcon = () => (
@@ -123,9 +124,9 @@ SelectField.propTypes = {
     disabled: propTypes.bool,
     filled: propTypes.bool,
     dense: propTypes.bool,
-    valid: propTypes.bool,
-    warning: propTypes.bool,
-    error: propTypes.bool,
+    valid: statusPropType,
+    warning: statusPropType,
+    error: statusPropType,
     loading: propTypes.bool,
     initialFocus: propTypes.bool,