+++ src/Switch/index.js
@@ -1,9 +1,9 @@
-import React, { Component, createRef } from 'react'
-import propTypes from 'prop-types'
 import cx from 'classnames'
+import propTypes from 'prop-types'
+import React, { Component, createRef } from 'react'
 
+import { statusPropType } from '../common-prop-types'
 import { SwitchIcon } from '../icons/Switch.js'
-
 import styles from './styles'
 
 const Input = React.forwardRef(
@@ -118,9 +118,9 @@ Switch.propTypes = {
     required: propTypes.bool,
     checked: propTypes.bool,
     disabled: propTypes.bool,
-    valid: propTypes.bool,
-    warning: propTypes.bool,
-    error: propTypes.bool,
+    valid: statusPropType,
+    warning: statusPropType,
+    error: statusPropType,
     initialFocus: propTypes.bool,
 
     onFocus: propTypes.func,