+++ src/Tabs/index.js
@@ -1,0 +1,3 @@
+export { TabBar } from './TabBar'
+export { Tab } from './Tab'
+export { ScrollBar } from './ScrollBar'